Optimizations Round1
split some var declarations and global util functions into a seperate
file, cleaned up the dode a bit and did a few optimizations.

This is just a run-up to a bigg-ass commit thats coming. I plan to
tableize the calcualtor and modify the function generation/ mesh build
system so that the entire  thing is a module and builds itself.

After this I will start working on a proper UI.